+++ src/main/java/farquaad/command/ExitCommand.java
@@ -1,0 +1,17 @@
+package farquaad.command;
+
+import farquaad.TaskList;
+import farquaad.storage.Storage;
+import farquaad.ui.Ui;
+
+public class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.displayGoodbye();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
