+++ src/main/java/farquaad/command/ToDoCommand.java
@@ -1,0 +1,29 @@
+package farquaad.command;
+
+import farquaad.farquaadexception.*;
+import farquaad.TaskList;
+import farquaad.storage.Storage;
+import farquaad.task.Task;
+import farquaad.ui.Ui;
+import java.io.IOException;
+
+public class ToDoCommand extends Command {
+    private String arguments;
+
+    public ToDoCommand(String arguments) {
+        this.arguments = arguments;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage)
+            throws FarquaadException, IOException {
+        if (arguments.trim().isEmpty()) {
+            throw new EmptyDescriptionException("todo");
+        }
+
+        Task todo = new Task.ToDo(arguments.trim());
+        tasks.add(todo);
+        ui.displayTaskAdded(todo, tasks.size());
+        storage.save(tasks.getTasks());
+    }
+}
