+++ src/main/java/farquaad/farquaadexception/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package farquaad.farquaadexception;
+
+public class EmptyDescriptionException extends FarquaadException {
+    public EmptyDescriptionException(String task) {
+        super("lmao the description of a " + task + " cannot be empty.");
+    }
+}
