+++ src/main/java/farquaad/farquaadexception/FarquaadException.java
@@ -1,0 +1,11 @@
+package farquaad.farquaadexception;
+
+public class FarquaadException extends Exception {
+    public FarquaadException(String msg) {
+        super(msg);
+    }
+}
+
+
+
+
