+++ src/main/java/farquaad/farquaadexception/InvalidIndexException.java
@@ -1,0 +1,7 @@
+package farquaad.farquaadexception;
+
+public class InvalidIndexException extends FarquaadException {
+    public InvalidIndexException() {
+        super("the farquaad.task number is invalid.");
+    }
+}
