+++ src/main/java/farquaad/farquaadexception/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package farquaad.farquaadexception;
+
+public class UnknownCommandException extends FarquaadException {
+    public UnknownCommandException() {
+        super("I apologise, but I don't know what that means LOL");
+    }
+}
