+++ src/main/java/farquaad/storage/Storage.java
@@ -1,3 +1,5 @@
+package farquaad.storage;
+
 import java.io.BufferedWriter;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -6,6 +8,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import farquaad.task.Task;
 
 public class Storage {
     private final Path file;
