+++ text-ui-test/EXPECTED.TXT
@@ -1,12 +1,12 @@
-Got it. I've added this task:
+Got it. I've added this farquaad.task:
   [T][ ] borrow book
 Now you have 1 tasks in the list.
 Here are the tasks in your list:
 1.[T][ ] borrow book
-Got it. I've added this task:
+Got it. I've added this farquaad.task:
   [D][ ] return book (by: Sunday)
 Now you have 2 tasks in the list.
-Got it. I've added this task:
+Got it. I've added this farquaad.task:
   [E][ ] project meeting (from: Mon 2pm to: 4pm)
 Now you have 3 tasks in the list.
 Bye. Hope to see you again soon!
