+++ text-ui-test/sources.txt
@@ -1,2 +1,2 @@
 "C:\Users\gohji\IdeaProjects\2103 chatbot\ip\src\main\java\farquaad.java"
-"C:\Users\gohji\IdeaProjects\2103 chatbot\ip\src\main\java\Task.java"
+"C:\Users\gohji\IdeaProjects\2103 chatbot\ip\src\main\java\farquaad.task.Task.java"
